+++ packages/riverbloc/CHANGELOG.md
@@ -1,4 +1,9 @@
-## [0.2.0] - 2021-03-07
+## [0.2.2] - 2021-03-18
+
+- Upgrading dependency to bloc 7.0.0 (stable)
+- Using the new `BlocBase` class.
+
+## [0.2.1] - 2021-03-07
 
 - Updating meta dependency